requirements.txt

langchain-community
chromadb
transformers
sentence-transformers
groq
